Also read commands from standard input.
This adds the ability to also read commands from standard input (if
standard input is not a terminal).

Also, force --quiet id stdout is not a terminal.

Example:

    mkfifo ~/.castnow.fifo
    castnow --address a.b.c.d --quiet < ~/.castnow.fifo &

    echo pause >> ~/.castnow.fifo

This makes castnow suitable for scripting, or binding to WM keys.

Also, the following ensures that castnow does not exit after the first
command:

    sleep 2000000000 >> ~/.castnow.fifo & disown